fixed knife range bug fixed recoil crash with high fps

+++ src/main/java/com/blix/sixsiege/item/custom/KnifeItem.java
@@ -17,7 +17,7 @@ public class KnifeItem extends Item {
         super(settings);
         this.attackDamage = (float)attackDamage;
         ImmutableMultimap.Builder<EntityAttribute, EntityAttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, "Weapon modifier", (double)this.attackDamage, EntityAttributeModifier.Operation.ADDITION));
+        builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, "Weapon modifier", this.attackDamage, EntityAttributeModifier.Operation.ADDITION));
         this.attributeModifiers = builder.build();
     }
